requirements.txt
Django>=4.2.0,<5.0.0
graphene-django>=3.0.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
python-decouple>=3.8